+++ src/MAWSC.cs
@@ -4,7 +4,7 @@
 // https://github.com/spectrum-health-systems/MAWSC)
 // Apache v2 (https://apache.org/licenses/LICENSE-2.0)
 // Copyright 2021-2022 A Pretty Cool Program
-// ================================================================= [ 220617.090730-devbuild ]
+// ================================================================= [ 220617.091527-devbuild ]
 
 // MAWSC.cs
 // Entry point for MAWSC.
